feat(encoding): add WASM binding via core split
Extracts the iconv-lite-compatible character-encoding wrapper into
`_encoding-core` — `encoding_rs`-backed (WHATWG-spec) with the
iconv-lite divergence handling (strict latin1, strict windows-1252,
raw UTF-16 LE/BE, `cp***`/`latin*`/`utf16*` aliases) preserved. The
napi crate is a thin shim; WASM exposes the same `encode`, `decode`,
`encodingExists` surface — `encoding_rs` (~100 KB gzipped) is well
under the 500 KB browser budget.
