+++ crates/_encoding-core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "amigo-encoding-core"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
+
+[dependencies]
+encoding_rs = "0.8"
