+++ crates/encoding/Cargo.toml
@@ -9,7 +9,7 @@ crate-type = ["cdylib"]
 [dependencies]
 napi = { workspace = true }
 napi-derive = { workspace = true }
-encoding_rs = "0.8"
+amigo-encoding-core = { path = "../_encoding-core" }
 
 [build-dependencies]
 napi-build = "2"
