+++ crates/encoding/package.json
@@ -4,12 +4,30 @@
   "description": "Rust-powered character-encoding conversion via encoding_rs",
   "main": "index.js",
   "types": "index.d.ts",
+  "browser": "./wasm/pkg/amigo_encoding_wasm.js",
+  "exports": {
+    ".": {
+      "browser": {
+        "types": "./wasm/pkg/amigo_encoding_wasm.d.ts",
+        "default": "./wasm/pkg/amigo_encoding_wasm.js"
+      },
+      "default": {
+        "types": "./index.d.ts",
+        "default": "./index.js"
+      }
+    }
+  },
   "files": [
     "index.js",
     "index.d.ts",
     "native.cjs",
     "native.d.ts",
-    "README.md"
+    "README.md",
+    "wasm/pkg/amigo_encoding_wasm.js",
+    "wasm/pkg/amigo_encoding_wasm_bg.js",
+    "wasm/pkg/amigo_encoding_wasm_bg.wasm",
+    "wasm/pkg/amigo_encoding_wasm.d.ts",
+    "wasm/pkg/amigo_encoding_wasm_bg.wasm.d.ts"
   ],
   "keywords": [
     "encoding",
@@ -35,10 +53,13 @@
     "artifacts": "napi artifacts",
     "build": "napi build --platform --release --js native.cjs --dts native.d.ts",
     "build:debug": "napi build --platform --js native.cjs --dts native.d.ts",
-    "prepublishOnly": "napi pre-publish --skip-optional-publish --no-gh-release",
+    "build:wasm": "cd wasm && wasm-pack build --target bundler --release --out-dir pkg",
+    "build:all": "pnpm build && pnpm build:wasm",
+    "prepublishOnly": "pnpm build:wasm && napi pre-publish --skip-optional-publish --no-gh-release",
     "test": "vitest run __test__",
     "bench": "vitest bench",
     "test:conformance": "vitest run __conformance__",
+    "test:wasm": "cd wasm && wasm-pack test --node",
     "test:all": "vitest run",
     "test:upstream": "vitest run __conformance__/upstream.spec.ts"
   },
@@ -66,6 +87,7 @@
     "vsJs": "0.61-32x",
     "parity": "100%",
     "status": "Alternative",
+    "targets": ["node", "browser"],
     "competitors": [
       "iconv-lite@0.6.3"
     ]
