+++ crates/encoding/wasm/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "amigo-encoding-wasm"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+amigo-encoding-core = { path = "../../_encoding-core" }
+wasm-bindgen = { workspace = true }
+
+[dev-dependencies]
+wasm-bindgen-test = { workspace = true }
+
+[package.metadata.wasm-pack.profile.release]
+wasm-opt = false
