+++ docs/readmes/encoding.html
@@ -15,5 +15,10 @@ encodingExists('shift_jis')                // true
 encodingExists('utf-42')                   // false
 </code></pre>
 <p>Supports the full WHATWG encoding set (UTF-8, Latin-1, Shift_JIS, GB18030, Big5, etc.).</p>
+<h2 id="install-for-the-browser">Install for the browser</h2>
+<p>The same <code>import</code> works in Angular, React, Vite, esbuild, and webpack ≥ 5 — the bundler picks the WASM build via the <code>browser</code> conditional export:</p>
+<pre><code class="language-ts">import { encode, decode } from '@amigo-labs/encoding'
+</code></pre>
+<p><code>encoding_rs</code> is ~100 KB gzipped — well under the 500 KB browser budget. All iconv-lite-compat aliases (<code>latin0</code>, <code>cp932</code>, <code>utf16le</code>, etc.) and the strict latin1/windows-1252 semantics are preserved.</p>
 <h2 id="parity">Parity</h2>
 <p>Tests in <a href="./__conformance__"><code>__conformance__/</code></a> run a representative subset of the upstream <code>iconv-lite</code> test suite against this implementation. See <a href="./__conformance__/divergences.md"><code>divergences.md</code></a> for documented differences — notably encodings <code>iconv-lite</code> supports that <code>encoding_rs</code> doesn’t.</p>
